backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# LangChain ecosystem
langchain==0.1.4
langchain-community==0.0.16

# Embeddings and transformers
sentence-transformers
transformers


# Vector database
faiss-cpu

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
python-magic==0.4.27

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dateutil==2.8.2
aiofiles==23.2.1

# Testing and evaluation
pytest==7.4.4
pytest-asyncio==0.23.3
scikit-learn==1.4.0

# Monitoring
python-json-logger==2.0.7
